Added documentation

+++ Game Engine/Engine/time.hpp
@@ -13,33 +13,41 @@
 #include "Math/siconstants.hpp"
 
 namespace Time {
+  ///Get current time in nanoseconds as an integer
   inline uint64_t getNanoI() {
     return std::chrono::high_resolution_clock::now().time_since_epoch().count();
   }
+  ///Get current time in microseconds as an integer
   inline uint64_t getMicroI() {
     return getNanoI() * Math::SI::NANO_MICRO;
   }
+  ///Get current time in milliseconds as an integer
   inline uint64_t getMilliI() {
     return getNanoI() * Math::SI::NANO_MILLI;
   }
+  ///Get current time in seconds as an integer
   inline uint64_t getSecI() {
     return getNanoI() * Math::SI::NANO_ONE;
   }
   
+  ///Get current time in nanoseconds as floating point number
   inline double getNanoF() {
     return std::chrono::high_resolution_clock::now().time_since_epoch().count();
   }
-  inline uint64_t getMicroF() {
+  ///Get current time in microseconds as floating point number
+  inline double getMicroF() {
     return getNanoF() * Math::SI::NANO_MICRO;
   }
+  ///Get current time in milliseconds as floating point number
   inline double getMilliF() {
     return getNanoF() * Math::SI::NANO_MILLI;
   }
+  ///Get current time in seconds as floating point number
   inline double getSecF() {
     return getNanoF() * Math::SI::NANO_ONE;
   }
   
-  //Get the unix timestamp
+  ///Get the Unix timestamp
   inline uint64_t getDate() {
     return std::chrono::system_clock::now().time_since_epoch().count() * Math::SI::MICRO_ONE;
   }
